Remove Sparse Dataset Dependancy
Remove metric calculations that require the use of a spare data set. Instead create a function that will take the tidy dataset (like the output from a database) and  temporarily create a sparse dataset for using the 'vegan' package to compute diversity and evenness indices. This then requires only 1 input data table.